backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.11",
    "langchain-text-splitters>=1.0.0",
    "langchain-openai>=1.0.3",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "asyncpg>=0.30.0",
    "pgvector>=0.4.1",
    "pymupdf>=1.26.5",
    "python-multipart>=0.0.20",
]
